+++ pkg/sql/q_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/nzoschke/codon/pkg/db"
-	"github.com/nzoschke/codon/pkg/models"
+	"github.com/nzoschke/codon/pkg/sql/models"
 	"github.com/nzoschke/codon/pkg/sql/q"
 	"github.com/stretchr/testify/assert"
 )
